requirements.txt
async-timeout==5.0.1
boto3==1.40.7
chardet==5.2.0
coreapi==2.3.3
django-allauth==65.3.1
django-celery-beat==2.8.1
django-cors-headers==4.6.0
django-filter==24.3
django-mptt==0.17.0
django-redis==5.4.0
djangorestframework-simplejwt==5.3.1
docker==7.1.0
drf-spectacular==0.27.0
drf-spectacular-sidecar==2025.8.1
drf-yasg==1.21.10
fido2==1.2.0
flower==2.0.1
ipykernel==6.29.5
minio==7.2.15
# mysqlclient==2.2.6
openai==1.81.0
openpyxl==3.1.5
oss2==2.19.1
pandas==2.2.3
pdfplumber==0.11.4
pipdeptree==2.26.1
pipenv==2024.4.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyPDF2==3.0.1
pypng==0.20220715.0
pytest-django==4.9.0
pytest-cov==6.0.0
pytest-html==4.1.1
pytest-xdist==3.6.1
python3-openid==3.2.0
qrcode==8.0
django-otp>=1.2.0
requests-oauthlib==2.0.0
spacy==3.8.6
torch==2.7.0
transformers==4.52.2
uWSGI==2.0.28
wheel==0.46.1
xlrd==2.0.1
python-dotenv==1.0.0
django-phonenumber-field>=7.2.0
phonenumbers>=8.13.0
aliyun-python-sdk-core>=2.15.0
aliyun-python-sdk-dysmsapi>=2.1.0
